move client initialization to init file
