+++ nodes/launcher/launcher_node.py
@@ -16,49 +16,15 @@ routes = PromptServer.instance.routes
 UI_DIR = pathlib.Path(__file__).parent.parent.parent / "ui"
 dev_server_process = None
 
-def get_npm_command():
-    """Get the npm command for the current platform"""
-    npm_cmd = shutil.which("npm")
-    if npm_cmd is None:
-        raise RuntimeError("npm not found in PATH. Please install Node.js and npm.")
-    return npm_cmd
-
-def run_npm_command(command, cwd):
-    """Run an npm command in a platform-agnostic way"""
-    npm_cmd = get_npm_command()
-    full_cmd = [npm_cmd] + command.split()
-    
-    try:
-        return subprocess.run(
-            full_cmd,
-            cwd=cwd,
-            check=True,
-            capture_output=True,
-            text=True
-        )
-    except subprocess.CalledProcessError as e:
-        print(f"Error running npm command: {e.stderr}")
-        raise
-
-def ensure_ui_ready():
-    """Ensure the UI is ready (dependencies installed)"""
-    try:
-        # Check if node_modules exists
-        if not (UI_DIR / "node_modules").exists():
-            print("[ComfyStream] Installing dependencies...")
-            run_npm_command("install", UI_DIR)
-            print("[ComfyStream] Dependencies installed successfully")
-    except Exception as e:
-        print(f"[ComfyStream] Error preparing UI: {str(e)}")
-        raise
-
 def start_dev_server():
     """Start the Next.js dev server"""
     global dev_server_process
     if dev_server_process is None:
         try:
             print("[ComfyStream] Starting dev server...")
-            npm_cmd = get_npm_command()
+            npm_cmd = shutil.which("npm")
+            if npm_cmd is None:
+                raise RuntimeError("npm not found in PATH")
             # Start the dev server on port 3000
             dev_server_process = subprocess.Popen(
                 [npm_cmd, "run", "dev"],
@@ -88,9 +54,6 @@ def cleanup_dev_server():
 
 # Register cleanup on exit
 atexit.register(cleanup_dev_server)
-
-# Ensure UI dependencies are installed on module load
-ensure_ui_ready()
 
 @routes.post('/launch_comfystream')
 async def launch_comfystream(request):
